web/feat: hide completed tasks button reposition
