+++ packages/web/app/_components/tasks/tasks-page.tsx
@@ -1,9 +1,7 @@
 'use client';
 
 import { FC, useState } from 'react';
-import { useTranslations } from 'next-intl';
-import { Button, Card, Group, Stack } from '@mantine/core';
-import { IconEye, IconEyeOff } from '@tabler/icons-react';
+import { Card, Group, Stack } from '@mantine/core';
 import { useProjectsTabs } from '@/hooks/use-projects-tabs';
 import { useTasks } from '@/hooks/use-tasks';
 import { ProjectsTabs } from '../projects-tabs';
@@ -14,16 +12,13 @@ import styles from '@/styles/components/tasks.module.scss';
 
 export const TasksPage: FC = () => {
   const {
-    t,
     projectId,
     projectName,
     tabs,
     todos,
     done,
-    hideCompleted,
     popoverOpen,
     setPopoverOpen,
-    handleHideCompleted,
     handleChange,
     onCreate,
     onEdit,
@@ -38,26 +33,12 @@ export const TasksPage: FC = () => {
 
   return (
     <Stack h="100%" gap={8}>
-      <Group justify="space-between">
+      <Group>
         <ProjectsTabs
           defaultTab={projectId}
           tabs={tabs}
           onChange={handleChange}
         />
-        <Button
-          size="sm"
-          variant="subtle"
-          bg="white"
-          c="neutral.7"
-          radius={6}
-          leftSection={
-            hideCompleted ? <IconEye size={20} /> : <IconEyeOff size={20} />
-          }
-          className={styles.hideButton}
-          onClick={handleHideCompleted}
-        >
-          {hideCompleted ? t('showAction') : t('hideAction')}
-        </Button>
       </Group>
       <Card h="100%" radius={24} className={styles.tasks}>
         <TasksHeader
@@ -71,7 +52,6 @@ export const TasksPage: FC = () => {
           done={done}
           onEdit={onEdit}
           onDelete={onDelete}
-          hideCompleted={hideCompleted}
         />
       </Card>
     </Stack>
@@ -79,8 +59,6 @@ export const TasksPage: FC = () => {
 };
 
 function useTasksPage() {
-  const t = useTranslations('tasks');
-  const [hideCompleted, setHideCompleted] = useState(false);
   const [popoverOpen, setPopoverOpen] = useState(false);
   const { selectedProject, tabs, handleChange } = useProjectsTabs();
   const { id: projectId, name: projectName } = selectedProject ?? {};
@@ -89,21 +67,14 @@ function useTasksPage() {
     () => setPopoverOpen(false)
   );
 
-  const handleHideCompleted = () => {
-    setHideCompleted(!hideCompleted);
-  };
-
   return {
-    t,
     projectId,
     projectName,
     tabs,
     todos,
     done,
-    hideCompleted,
     popoverOpen,
     setPopoverOpen,
-    handleHideCompleted,
     handleChange,
     onCreate,
     onEdit,
