+++ packages/web/app/_components/tasks/tasks-wrapper.tsx
@@ -1,22 +1,31 @@
-import { FC } from 'react';
+import { FC, useState } from 'react';
 import { useTranslations } from 'next-intl';
-import { Box, Flex, Stack } from '@mantine/core';
+import { Box, Button, Flex, Stack } from '@mantine/core';
+import { IconEye, IconEyeOff } from '@tabler/icons-react';
 import { Task } from '@/domain/queries/tasks-query';
 import { TasksList } from './tasks-list';
-import overflowStyles from '@/styles/utils/overflow.module.scss';
 import { TasksEmptyPlaceholder } from './tasks-empty-placeholder';
+import overflowStyles from '@/styles/utils/overflow.module.scss';
+import styles from '@/styles/components/tasks.module.scss';
 
 type Props = {
   todos: Task[];
   done: Task[];
   onEdit: (task: Task) => Promise<void>;
   onDelete: (id: string) => Promise<void>;
-  hideCompleted: boolean;
 };
 
 export const TasksWrapper: FC<Props> = (props) => {
-  const { t, isEmpty, todos, done, onEdit, onDelete, hideCompleted } =
-    useTasksWrapper(props);
+  const {
+    t,
+    isEmpty,
+    todos,
+    done,
+    onEdit,
+    onDelete,
+    hideCompleted,
+    handleHideCompleted,
+  } = useTasksWrapper(props);
 
   return (
     <Box h="100%" className={overflowStyles['overflow-auto']}>
@@ -34,29 +43,54 @@ export const TasksWrapper: FC<Props> = (props) => {
               onDelete={onDelete}
             />
           )}
-          {!hideCompleted && !!done.length && (
-            <TasksList
-              title={t('done')}
-              tasks={done}
-              onEdit={onEdit}
-              onDelete={onDelete}
-            />
-          )}
+          <Stack>
+            <Button
+              size="sm"
+              variant="subtle"
+              bg="white"
+              c="neutral.7"
+              radius={6}
+              leftSection={
+                hideCompleted ? <IconEye size={20} /> : <IconEyeOff size={20} />
+              }
+              className={styles.hideButton}
+              onClick={handleHideCompleted}
+            >
+              {hideCompleted ? t('showAction') : t('hideAction')}
+            </Button>
+            {!hideCompleted && !!done.length && (
+              <TasksList
+                title={t('done')}
+                tasks={done}
+                onEdit={onEdit}
+                onDelete={onDelete}
+              />
+            )}
+          </Stack>
         </Stack>
       )}
     </Box>
   );
 };
 
-function useTasksWrapper({
-  todos,
-  done,
-  onEdit,
-  onDelete,
-  hideCompleted,
-}: Props) {
+function useTasksWrapper({ todos, done, onEdit, onDelete }: Props) {
   const t = useTranslations('tasks');
+  const [hideCompleted, setHideCompleted] = useState(false);
+
   const isEmpty = !todos.length && !done.length;
 
-  return { t, isEmpty, todos, done, onEdit, onDelete, hideCompleted };
+  const handleHideCompleted = () => {
+    setHideCompleted(!hideCompleted);
+  };
+
+  return {
+    t,
+    isEmpty,
+    todos,
+    done,
+    onEdit,
+    onDelete,
+    hideCompleted,
+    handleHideCompleted,
+  };
 }
