+++ packages/web/styles/components/tasks.module.scss
@@ -11,6 +11,7 @@
 }
 
 .hideButton {
+  align-self: flex-start;
   border: 1px solid var(--mantine-color-neutral-3);
 }
